+++ R/EmpiricalCalibrationUsingAsymptotics.R
@@ -66,6 +66,11 @@ fitNull <- function(logRr, seLogRr) {
     seLogRr <- seLogRr[!is.na(logRr)]
     logRr <- logRr[!is.na(logRr)]
   }
+  if (any(abs(logRr) > log(100))) {
+    warning("Estimate(s) with extreme logRr detected: abs(logRr) > log(100). Removing before fitting null distribution")
+    seLogRr <- seLogRr[abs(logRr) <= log(100)]
+    logRr <- logRr[abs(logRr) <= log(100)]
+  }
   if (length(logRr) == 0) {
     warning("No estimates remaining")
     null <- c(NA, NA)