+++ R/EmpiricalCalibrationUsingMcmc.R
@@ -177,6 +177,11 @@ fitMcmcNull <- function(logRr, seLogRr, iter = 100000) {
     seLogRr <- seLogRr[!is.na(logRr)]
     logRr <- logRr[!is.na(logRr)]
   }
+  if (any(abs(logRr) > log(100))) {
+    warning("Estimate(s) with extreme logRr detected: abs(logRr) > log(100). Removing before fitting null distribution")
+    seLogRr <- seLogRr[abs(logRr) <= log(100)]
+    logRr <- logRr[abs(logRr) <= log(100)]
+  }
   if (length(logRr) == 0) {
     warning("No valid estimates left. Returning undefined null distribution")
     mcmc <- list(chain = matrix(c(NA,NA), nrow = 1, ncol = 2))